Added camera/projection frustum visualization

+++ Timer.cpp
@@ -5,15 +5,15 @@ Timer::Timer()
 {
 }
 
-float Timer::Mark()
+double Timer::Mark()
 {
 	const auto _old = std::move(m_Last);
 	m_Last = std::chrono::steady_clock::now();
-	const std::chrono::duration<float> _frameTime = m_Last - _old;
+	const std::chrono::duration<double> _frameTime = m_Last - _old;
 	return _frameTime.count();
 }
 
-float Timer::Peek() const
+double Timer::Peek() const
 {
-	return std::chrono::duration<float>(std::chrono::steady_clock::now() - m_Last).count();
+	return std::chrono::duration<double>(std::chrono::steady_clock::now() - m_Last).count();
 }
